LibCargaCreada
Ya espacio para guardar la partida creada y reorganizacion del codigo
