+++ JuegoRPG-C--main/protecyoFinal/Objeto.h
@@ -4,7 +4,6 @@
 #include <stdio.h>
 #include <stdlib.h>
 #define PERGAMINO "pergamino.bin"
-#define POCION "pocion.bin"
 #define ARMA "arma.bin"
 #define ARMADURA "armadura.bin"
 
@@ -17,5 +16,13 @@ typedef struct Objeto
 } Objeto;
 
 Objeto crearobjeto(); ///Crea un objeto
+Objeto dropoti();
+void persistenciaobj(Objeto elobjeto); /// Carga a un archivo un objeto
+void persistirpergamino(Objeto elobjeto);
+void persistirarma(Objeto elobjeto);
+void persistirarmadura(Objeto elobjeto);
+Objeto depersistirpergamino(int ultimo);///Ultimo es un entero para reposicionar luego el puntero del archivo
+Objeto depersistirarma(int ultimo);
+Objeto depersistirarmadura(int ultimo);
 
 #endif // OBJETO_H_INCLUDED
