+++ JuegoRPG-C--main/protecyoFinal/libreriaNueva.h
@@ -6,14 +6,5 @@
 #include <windows.h>
 #include "Codito2.h"
 #include "Jugador.h"
-typedef struct mob
-{
-    char nombre[30];
-    char prefacio[200];
-    int vida;
-    int danio;
-    int estado;///0 muerto, 1 vivo
-}mob;
 
-nododoble*recogerLootCofre(Heroe*jugador, nododoble*lista);
 #endif // LIBRERIANUEVA_H_INCLUDED
