--- libreriaNueva.c
@@ -1,4 +1,0 @@
-#include "libreriaNueva.h"
-#include "Jugador.h"
-#include "objeto.h"
-
