--- libreriaNueva.h
@@ -1,10 +1,0 @@
-#ifndef LIBRERIANUEVA_H_INCLUDED
-#define LIBRERIANUEVA_H_INCLUDED
-#include <string.h>
-#include <stdio.h>
-#include <stdlib.h>
-#include <windows.h>
-#include "Codito2.h"
-#include "Jugador.h"
-
-#endif // LIBRERIANUEVA_H_INCLUDED
