--- mapa.h
@@ -1,34 +1,0 @@
-#ifndef ARBOL_H_INCLUDED
-#define ARBOL_H_INCLUDED
-#include <stdio.h>
-#include <stdlib.h>
-#include "Jugador.h"
-#include "libreriaNueva.h"
-
-typedef struct nodoArbol
-{
-    int id; ///del 1 al 7, inicia en 4
-    int estado; ///1 completado, 0 no completado
-    int encuentro; ///1 si es combate, 0 si es cofre
-    struct nodoArbol* derecha; //apunta a derecha
-    struct nodoArbol* izquierda;//apunta a izquierda
-} nodoArbol;
-
-nodoArbol* inicArbol();
-nodoArbol* crearNodoArbol(int dato);
-nodoArbol* insertar(nodoArbol* arbol,int dato);
-void preorder(nodoArbol* arbol);
-void inorder(nodoArbol* arbol);
-void postorder(nodoArbol* arbol);
-nodoArbol* buscarArbol(nodoArbol* arbol, int dato);
-nodoArbol* NMI(nodoArbol* nodo);
-nodoArbol* NMD(nodoArbol* nodo);
-nodoArbol* borrarNodo(nodoArbol* nodo, int dato);
-int esHoja (nodoArbol* nodo);
-// prototipado manejo de mapa
-nodoArbol*cargaMapa(nodoArbol*arbol);
-void avanzarIzquierda(nodoArbol*mapa, int habitacionActual, Heroe*jugador);
-void avanzarDerecha(nodoArbol*mapa, int habitacionActual, Heroe*jugador);
-// fin prototipado manejo de mapa
-
-#endif // ARBOL_H_INCLUDED
